# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: SNAX Supplier Import API
  version: 1.1.0
  description: API внешнего сервиса нормализации прайс-листов и обмена пакетами с 1С:УТ. Контракт не предоставляет расчёт
    потребности и создание заказа поставщику. Дополнение 1.1 добавляет документы поставки и receipt-package для приемки в
    1С.
servers:
- url: https://snax-import.example/api/v1
  description: Production placeholder
security:
- bearerAuth: []
paths:
  /imports:
    post:
      operationId: createImport
      summary: Загрузить исходный файл
      parameters:
      - $ref: '#/components/parameters/CorrelationId'
      - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
              - file
              properties:
                file:
                  type: string
                  format: binary
                supplierCode:
                  type:
                  - string
                  - 'null'
                profileCode:
                  type:
                  - string
                  - 'null'
      responses:
        '202':
          description: Файл зарегистрирован, обработка поставлена в очередь
          headers:
            X-Correlation-ID:
              $ref: '#/components/headers/CorrelationId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ImportAccepted'
        '409':
          description: Точная копия уже зарегистрирована
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
        default:
          $ref: '#/components/responses/ProblemResponse'
  /imports/{importId}:
    get:
      operationId: getImport
      summary: Получить состояние импорта
      parameters:
      - $ref: '#/components/parameters/ImportId'
      - $ref: '#/components/parameters/CorrelationId'
      responses:
        '200':
          description: Состояние импорта
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ImportStatus'
        default:
          $ref: '#/components/responses/ProblemResponse'
  /imports/{importId}/rows:
    get:
      operationId: listImportRows
      summary: Просмотреть строки импорта
      parameters:
      - $ref: '#/components/parameters/ImportId'
      - $ref: '#/components/parameters/Cursor'
      - name: rowType
        in: query
        schema:
          type:
          - string
          - 'null'
      - name: qualityStatus
        in: query
        schema:
          type:
          - string
          - 'null'
      responses:
        '200':
          description: Страница строк
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RowPage'
        default:
          $ref: '#/components/responses/ProblemResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    CorrelationId:
      name: X-Correlation-ID
      in: header
      required: false
      schema:
        type: string
        maxLength: 100
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema:
        type: string
        minLength: 16
        maxLength: 200
    ImportId:
      name: importId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    Cursor:
      name: cursor
      in: query
      required: false
      schema:
        type:
        - string
        - 'null'
...
  headers:
    CorrelationId:
      schema:
        type: string
  responses:
    ProblemResponse:
      description: Ошибка
      headers:
        X-Correlation-ID:
          $ref: '#/components/headers/CorrelationId'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Problem'
  schemas:
    Problem:
      type: object
      additionalProperties: false
      required:
      - code
      - message
      - retryable
      - correlationId
      properties:
        code:
          type: string
          pattern: ^[A-Z0-9_]+$
        message:
          type: string
        details:
          type:
          - object
          - 'null'
        field:
          type:
          - string
          - 'null'
        retryable:
          type: boolean
        correlationId:
          type: string
    ImportAccepted:
      type: object
      additionalProperties: false
      required:
      - importId
      - status
      - statusUrl
      properties:
        importId:
          type: string
          format: uuid
        status:
          type: string
          enum:
          - RECEIVED
          - STORED
          - QUEUED
          - PROCESSING
          - READY_FOR_REVIEW
          - FAILED
          - CANCELLED
          - DUPLICATE
        statusUrl:
          type: string
          format: uri-reference
    ImportStatus:
      type: object
      additionalProperties: false
      required:
      - importId
      - status
      - createdAt
      - summary
      properties:
        importId:
          type: string
          format: uuid
        status:
          type: string
          enum:
          - RECEIVED
          - STORED
          - QUEUED
          - DUPLICATE
          - PROCESSING
          - READY_FOR_REVIEW
          - FAILED
          - CANCELLED
          - PROFILE_REVIEW
          - VALIDATED
          - BLOCKED
          - READY_FOR_1C
          - DELIVERED
          - ACCEPTED_BY_1C
          - REJECTED_BY_1C
          - ERROR
          - ARCHIVED
        createdAt:
          type: string
          format: date-time
        profileCode:
          type:
          - string
          - 'null'
        profileVersion:
          type:
          - string
          - 'null'
        summary:
          type: object
          additionalProperties: true
...
    RowPage:
      type: object
      additionalProperties: false
      required:
      - items
      - nextCursor
      properties:
        items:
          type: array
          items:
            type: object
            additionalProperties: true
        nextCursor:
          type:
          - string
          - 'null'
...